ssh/tailssh: remove init funcs, allow running with tsnet
It was using an init func to register itself with ipnlocal,
move that to be an explicit registration in `cmd/tailscaled`.

The SSH server required a parameter for the incubatorPath to be provided
and if it wasn't it would exec the command directly. This was used by
tests, instead make the path an optional argument and add another field
set only by tests which allows direct command execution.

This turns tailssh into a library that can be embedded in applications
using tsnet. Add an example for setup too.

Updates #1748

Signed-off-by: Maisem Ali <maisem@tailscale.com>